Monster.java - enhancing feeder exp MonsterHealingForUser.java - health progress is a float MonsterStuff.proto - MonsterProto has enhancingFeederExp, healthProgress is float not int, and UserEvolutionProto is UserMonsterEvolutionProto MonsterRetrieveUtils.java - extracting enhancingFeeder exp MonsterStuffUtils.java - (monster healing for user) health progress is a float CreateInfoProtoUtils.java - setting enhance feeder exp and UserEvolutionProto is renamed to UserMonsterEvolutionProto
EventMonster.proto - UserEvolutionProto is now renamed to UserMonsterEvolutionProto
EventStartup.proto - above
EvolveMonsterController.java - above
StartupController.java - above

UpdateUtils.java - (MonsterHealingForUser) healthProgress is a float

+++ src/main/java/com/lvl6/server/controller/utils/MonsterStuffUtils.java
@@ -153,7 +153,7 @@ public class MonsterStuffUtils {
   		}
   		
 //  		int userStructHospitalId = umhp.getUserHospitalStructId();
-  		int healthProgress = umhp.getHealthProgress();
+  		float healthProgress = umhp.getHealthProgress();
   		int priority = umhp.getPriority();
 //  		Date queuedTime = new Date(umhp.getQueuedTimeMillis());
   		MonsterHealingForUser mhfu = new MonsterHealingForUser(userId, monsterForUserId,
