+++ src/main/java/com/lvl6/utils/utilmethods/UpdateUtils.java
@@ -679,7 +679,7 @@ public class UpdateUtils implements UpdateUtil {
 			
 //			int userStructHospitalId = mhfu.getUserStructHospitalId();
 //			aRow.put(DBConstants.MONSTER_HEALING_FOR_USER__USER_STRUCT_HOSPITAL_ID, userStructHospitalId);
-			int healthProgress = mhfu.getHealthProgress();
+			float healthProgress = mhfu.getHealthProgress();
 			aRow.put(DBConstants.MONSTER_HEALING_FOR_USER__HEALTH_PROGRESS, healthProgress);
 			int priority = mhfu.getPriority();
 			aRow.put(DBConstants.MONSTER_HEALING_FOR_USER__PRIORITY, priority);
